refactor: moved additional fields to insertcommon
