+++ go/vt/vtgate/engine/insert_common.go
@@ -1,0 +1,232 @@
+/*
+Copyright 2023 The Vitess Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package engine
+
+import (
+	"context"
+	"fmt"
+
+	"vitess.io/vitess/go/sqltypes"
+	"vitess.io/vitess/go/vt/key"
+	querypb "vitess.io/vitess/go/vt/proto/query"
+	"vitess.io/vitess/go/vt/vterrors"
+	"vitess.io/vitess/go/vt/vtgate/vindexes"
+)
+
+type InsertCommon struct {
+	// Opcode is the execution opcode.
+	Opcode InsertOpcode
+
+	// Keyspace specifies the keyspace to send the query to.
+	Keyspace *vindexes.Keyspace
+
+	// Ignore is for INSERT IGNORE and INSERT...ON DUPLICATE KEY constructs
+	// for sharded cases.
+	Ignore bool
+
+	// TableName is the name of the table on which row will be inserted.
+	TableName string
+
+	// Option to override the standard behavior and allow a multi-shard insert
+	// to use single round trip autocommit.
+	//
+	// This is a clear violation of the SQL semantics since it means the statement
+	// is not atomic in the presence of PK conflicts on one shard and not another.
+	// However, some application use cases would prefer that the statement partially
+	// succeed in order to get the performance benefits of autocommit.
+	MultiShardAutocommit bool
+
+	// QueryTimeout contains the optional timeout (in milliseconds) to apply to this query
+	QueryTimeout int
+
+	// ForceNonStreaming is true when the insert table and select table are same.
+	// This will avoid locking by the select table.
+	ForceNonStreaming bool
+
+	PreventAutoCommit bool
+
+	// ColVindexes are the vindexes that will use the VindexValues
+	ColVindexes []*vindexes.ColumnVindex
+
+	// Insert needs tx handling
+	txNeeded
+}
+
+type ksID = []byte
+
+// GetKeyspaceName specifies the Keyspace that this primitive routes to.
+func (ic *InsertCommon) GetKeyspaceName() string {
+	return ic.Keyspace.Name
+}
+
+// GetTableName specifies the table that this primitive routes to.
+func (ic *InsertCommon) GetTableName() string {
+	return ic.TableName
+}
+
+// GetFields fetches the field info.
+func (ic *InsertCommon) GetFields(context.Context, VCursor, map[string]*querypb.BindVariable) (*sqltypes.Result, error) {
+	return nil, vterrors.VT13001("unexpected fields call for insert query")
+}
+
+// processPrimary maps the primary vindex values to the keyspace ids.
+func (ic *InsertCommon) processPrimary(ctx context.Context, vcursor VCursor, vindexColumnsKeys []sqltypes.Row, colVindex *vindexes.ColumnVindex) ([]ksID, error) {
+	destinations, err := vindexes.Map(ctx, colVindex.Vindex, vcursor, vindexColumnsKeys)
+	if err != nil {
+		return nil, err
+	}
+
+	keyspaceIDs := make([]ksID, len(destinations))
+	for i, destination := range destinations {
+		switch d := destination.(type) {
+		case key.DestinationKeyspaceID:
+			// This is a single keyspace id, we're good.
+			keyspaceIDs[i] = d
+		case key.DestinationNone:
+			// No valid keyspace id, we may return an error.
+			if !ic.Ignore {
+				return nil, fmt.Errorf("could not map %v to a keyspace id", vindexColumnsKeys[i])
+			}
+		default:
+			return nil, fmt.Errorf("could not map %v to a unique keyspace id: %v", vindexColumnsKeys[i], destination)
+		}
+	}
+
+	return keyspaceIDs, nil
+}
+
+// processOwned creates vindex entries for the values of an owned column.
+func (ic *InsertCommon) processOwned(ctx context.Context, vcursor VCursor, vindexColumnsKeys []sqltypes.Row, colVindex *vindexes.ColumnVindex, ksids []ksID) error {
+	if !ic.Ignore {
+		return colVindex.Vindex.(vindexes.Lookup).Create(ctx, vcursor, vindexColumnsKeys, ksids, false /* ignoreMode */)
+	}
+
+	// InsertIgnore
+	var createIndexes []int
+	var createKeys []sqltypes.Row
+	var createKsids []ksID
+
+	for rowNum, rowColumnKeys := range vindexColumnsKeys {
+		if ksids[rowNum] == nil {
+			continue
+		}
+		createIndexes = append(createIndexes, rowNum)
+		createKeys = append(createKeys, rowColumnKeys)
+		createKsids = append(createKsids, ksids[rowNum])
+	}
+	if createKeys == nil {
+		return nil
+	}
+
+	err := colVindex.Vindex.(vindexes.Lookup).Create(ctx, vcursor, createKeys, createKsids, true)
+	if err != nil {
+		return err
+	}
+	// After creation, verify that the keys map to the keyspace ids. If not, remove
+	// those that don't map.
+	verified, err := vindexes.Verify(ctx, colVindex.Vindex, vcursor, createKeys, createKsids)
+	if err != nil {
+		return err
+	}
+	for i, v := range verified {
+		if !v {
+			ksids[createIndexes[i]] = nil
+		}
+	}
+	return nil
+}
+
+// processUnowned either reverse maps or validates the values for an unowned column.
+func (ic *InsertCommon) processUnowned(ctx context.Context, vcursor VCursor, vindexColumnsKeys []sqltypes.Row, colVindex *vindexes.ColumnVindex, ksids []ksID) error {
+	var reverseIndexes []int
+	var reverseKsids []ksID
+
+	var verifyIndexes []int
+	var verifyKeys []sqltypes.Row
+	var verifyKsids []ksID
+
+	// Check if this VIndex is reversible or not.
+	reversibleVindex, isReversible := colVindex.Vindex.(vindexes.Reversible)
+
+	for rowNum, rowColumnKeys := range vindexColumnsKeys {
+		// If we weren't able to determine a keyspace id from the primary VIndex, skip this row
+		if ksids[rowNum] == nil {
+			continue
+		}
+
+		if rowColumnKeys[0].IsNull() {
+			// If the value of the column is `NULL`, but this is a reversible VIndex,
+			// we will try to generate the value from the keyspace id generated by the primary VIndex.
+			if isReversible {
+				reverseIndexes = append(reverseIndexes, rowNum)
+				reverseKsids = append(reverseKsids, ksids[rowNum])
+			}
+
+			// Otherwise, don't do anything. Whether `NULL` is a valid value for this column will be
+			// handled by MySQL.
+		} else {
+			// If a value for this column was specified, the keyspace id values from the
+			// secondary VIndex need to be verified against the keyspace id from the primary VIndex
+			verifyIndexes = append(verifyIndexes, rowNum)
+			verifyKeys = append(verifyKeys, rowColumnKeys)
+			verifyKsids = append(verifyKsids, ksids[rowNum])
+		}
+	}
+
+	// Reverse map values for secondary VIndex columns from the primary VIndex's keyspace id.
+	if reverseKsids != nil {
+		reverseKeys, err := reversibleVindex.ReverseMap(vcursor, reverseKsids)
+		if err != nil {
+			return err
+		}
+
+		for i, reverseKey := range reverseKeys {
+			// Fill the first column with the reverse-mapped value.
+			vindexColumnsKeys[reverseIndexes[i]][0] = reverseKey
+		}
+	}
+
+	// Verify that the keyspace ids generated by the primary and secondary VIndexes match
+	if verifyIndexes != nil {
+		// If values were supplied, we validate against keyspace id.
+		verified, err := vindexes.Verify(ctx, colVindex.Vindex, vcursor, verifyKeys, verifyKsids)
+		if err != nil {
+			return err
+		}
+
+		var mismatchVindexKeys []sqltypes.Row
+		for i, v := range verified {
+			rowNum := verifyIndexes[i]
+			if !v {
+				if !ic.Ignore {
+					mismatchVindexKeys = append(mismatchVindexKeys, vindexColumnsKeys[rowNum])
+					continue
+				}
+
+				// Skip the whole row if this is a `INSERT IGNORE` or `INSERT ... ON DUPLICATE KEY ...` statement
+				// but the keyspace ids didn't match.
+				ksids[verifyIndexes[i]] = nil
+			}
+		}
+
+		if mismatchVindexKeys != nil {
+			return fmt.Errorf("values %v for column %v does not map to keyspace ids", mismatchVindexKeys, colVindex.Columns)
+		}
+	}
+
+	return nil
+}
