+++ go/vt/vtgate/planbuilder/operator_transformers.go
@@ -112,13 +112,13 @@ func transformInsertionSelection(ctx *plancontext.PlanningContext, op *operators
 	ins := dmlOp.(*operators.Insert)
 	eins := &engine.InsertSelect{
 		InsertCommon: &engine.InsertCommon{
-			Keyspace:  rb.Routing.Keyspace(),
-			TableName: ins.VTable.Name.String(),
-			Ignore:    ins.Ignore,
+			Keyspace:          rb.Routing.Keyspace(),
+			TableName:         ins.VTable.Name.String(),
+			Ignore:            ins.Ignore,
+			ForceNonStreaming: op.ForceNonStreaming,
+			ColVindexes:       ins.ColVindexes,
 		},
 		InsertRows:        engine.NewInsertRows(autoIncGenerate(ins.AutoIncrement)),
-		ForceNonStreaming: op.ForceNonStreaming,
-		ColVindexes:       ins.ColVindexes,
 		VindexValueOffset: ins.VindexValueOffset,
 	}
 	lp := &insert{eInsertSelect: eins}
@@ -548,15 +548,22 @@ func buildInsertLogicalPlan(
 	hints *queryHints,
 ) (logicalPlan, error) {
 	ins := op.(*operators.Insert)
+
+	ic := &engine.InsertCommon{
+		Opcode:      mapToInsertOpCode(rb.Routing.OpCode()),
+		Keyspace:    rb.Routing.Keyspace(),
+		TableName:   ins.VTable.Name.String(),
+		Ignore:      ins.Ignore,
+		ColVindexes: ins.ColVindexes,
+	}
+	if hints != nil {
+		ic.MultiShardAutocommit = hints.multiShardAutocommit
+		ic.QueryTimeout = hints.queryTimeout
+	}
+
 	eins := &engine.Insert{
-		InsertCommon: &engine.InsertCommon{
-			Opcode:    mapToInsertOpCode(rb.Routing.OpCode(), false),
-			Keyspace:  rb.Routing.Keyspace(),
-			TableName: ins.VTable.Name.String(),
-			Ignore:    ins.Ignore,
-		},
+		InsertCommon: ic,
 		InsertRows:   engine.NewInsertRows(autoIncGenerate(ins.AutoIncrement)),
-		ColVindexes:  ins.ColVindexes,
 		VindexValues: ins.VindexValues,
 	}
 	lp := &insert{eInsert: eins}
@@ -567,16 +574,11 @@ func buildInsertLogicalPlan(
 		eins.Prefix, eins.Mid, eins.Suffix = generateInsertShardedQuery(ins.AST)
 	}
 
-	if hints != nil {
-		eins.MultiShardAutocommit = hints.multiShardAutocommit
-		eins.QueryTimeout = hints.queryTimeout
-	}
-
 	eins.Query = generateQuery(stmt)
 	return lp, nil
 }
 
-func mapToInsertOpCode(code engine.Opcode, insertSelect bool) engine.InsertOpcode {
+func mapToInsertOpCode(code engine.Opcode) engine.InsertOpcode {
 	if code == engine.Unsharded {
 		return engine.InsertUnsharded
 	}
